% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baseline_measurements}
\alias{baseline_measurements}
\title{Baseline Measurements Dataset}
\format{
A data frame with 171163 rows and 18 variables.
}
\usage{
baseline_measurements
}
\description{
Summaries of demographic & baseline measures collected by arm or comparison group
and for the entire population of participants in the clinical study.
}
\details{
This illustrative data only includes participants with conditions "oncology", "malignant neoplasm", and "cancer" based on data extracted on 2022-06-07.

Definition of each variable can be found in the AACT Data Elements table on the AACT Data Dictionary page:
\url{https://aact.ctti-clinicaltrials.org/data_dictionary}
}
\keyword{datasets}
