% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brief_summaries}
\alias{brief_summaries}
\title{Brief Summaries Dataset}
\format{
A data frame with 42218 rows and 3 variables.
}
\usage{
brief_summaries
}
\description{
A single text column that provides a brief description of the study.
}
\details{
This illustrative data only includes participants with conditions "oncology", "malignant neoplasm", and "cancer" based on data extracted on 2022-06-07.

Definition of each variable can be found in the AACT Data Elements table on the AACT Data Dictionary page:
\url{https://aact.ctti-clinicaltrials.org/data_dictionary}
}
\keyword{datasets}
