% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{browse_interventions}
\alias{browse_interventions}
\title{Browse Interventions Dataset}
\format{
A data frame with 185518 rows and 5 variables.
}
\usage{
browse_interventions
}
\description{
NLM uses an internal algorithm to assess the data entered for a study and creates a list of standard
MeSH terms that describe the intervention(s) being addressed by the clinical trial. This table provides the results of NLM's assessment.
}
\details{
This illustrative data only includes participants with conditions "oncology", "malignant neoplasm", and "cancer" based on data extracted on 2022-06-07.

Definition of each variable can be found in the AACT Data Elements table on the AACT Data Dictionary page:
\url{https://aact.ctti-clinicaltrials.org/data_dictionary}
}
\keyword{datasets}
