% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conditions}
\alias{conditions}
\title{Conditions Dataset}
\format{
A data frame with 100806 row and 4 variables.
}
\usage{
conditions
}
\description{
Name(s) of the disease(s) or condition(s) studied in the clinical study, or the focus of the clinical study.
Can include NLM's Medical Subject Heading (MeSH)-controlled vocabulary terms.
}
\details{
This illustrative data only includes participants with conditions "oncology", "malignant neoplasm", and "cancer" based on data extracted on 2022-06-07.

Definition of each variable can be found in the AACT Data Elements table on the AACT Data Dictionary page:
\url{https://aact.ctti-clinicaltrials.org/data_dictionary}
}
\keyword{datasets}
