% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{designs}
\alias{designs}
\title{Designs Dataset}
\format{
A data frame with 41904 rows and 14 variables.
}
\usage{
designs
}
\description{
Description of how the study will be conducted,
including comparison group design and strategies for masking and allocating participants.
}
\details{
This illustrative data only includes participants with conditions "oncology", "malignant neoplasm", and "cancer" based on data extracted on 2022-06-07.

Definition of each variable can be found in the AACT Data Elements table on the AACT Data Dictionary page:
\url{https://aact.ctti-clinicaltrials.org/data_dictionary}
}
\keyword{datasets}
