% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_aact.R
\name{extract_aact}
\alias{extract_aact}
\title{extract_aact: Extract and create analysis AACT data}
\usage{
extract_aact(
  con = NULL,
  path_raw = NULL,
  file_type = "txt",
  board = NULL,
  terms_int,
  terms_cond,
  date_range = NULL,
  related_terms_int = NULL,
  related_terms_cond = NULL,
  outcome_type = NULL,
  outcome_list = NULL,
  outcome_name = NULL,
  is_shiny = FALSE,
  database = NULL,
  verbose = TRUE
)
}
\arguments{
\item{con}{Connection information for AACT (needed if connecting to database)}

\item{path_raw}{File path for raw AACT data (will ignore "con" argument if provided)}

\item{file_type}{File type for the raw AACT (ex: "txt" or "rda")}

\item{board}{Pins board}

\item{terms_int}{Intervention / treatment terms to search for (ex: paclitaxel). Set to NULL if no search.}

\item{terms_cond}{Search terms for condition(s) of interest (ex: non small lung cancer). Set to NULL if no search.}

\item{date_range}{Date range of interest (ex: TBD). Set to NULL if no search.}

\item{related_terms_int}{Related terms for intervention(s) (terms_int).
Default=NULL, which scrapes clinicaltrials.gov for term(s) related to terms_int}

\item{related_terms_cond}{Related terms for condition(s) (terms_cond).
Default=NULL, which scrapes clinicaltrials.gov for term(s) related to terms_cond}

\item{outcome_type}{Pre-specified outcomes (TBD; not currently available)}

\item{outcome_list}{List of outcome-based parameters to search for}

\item{outcome_name}{Vector of names for the searched outcomes (if NULL, uses generic names).}

\item{is_shiny}{Set to TRUE if used within a SHINY app. Default=FALSE}

\item{database}{Only used if con is not NULL. This is only needed if AACT data-tables are
stored such that data-tables are extracted using "select * from database.baseline" etc.}

\item{verbose}{If TRUE, messages describing process will be printed}
}
\value{
AACT analysis data(s) with treatments of interest and outcomes of interest
}
\description{
This function performs multiple steps: (1) Querying treatment and/or condition,
(2) Based on the identified trials for the matched treatment and/or condition, query
the outcome data. Key outputs include study-level information, along with outcome data
(treatment-specific, ex: median OS; and treatment-comparisons, ex: Hazard Ratios)
}
\details{
Add more details about extract_aact here
}
\examples{

\donttest{

library(aactreveal)

# Load package RDA for illustrative purposes #
table_names <- c("browse_interventions", "conditions", "design_group_interventions",
"design_groups","designs", "interventions", "outcome_analyses", "outcome_analysis_groups",
"outcome_measurements", "result_groups", "sponsors", "studies")

data(list=table_names, package="aactreveal")

terms_int <- "pembrolizumab"
terms_cond <- "breast cancer"
orr_list <- list(include = c("overall response rate", "objective response rate"),
exclude = c("disease control rate"),type = c("number"),range = c(0, 100) )
os_list <- list(include = c("overall survival"))
outcome_list <- list(orr_list, os_list)
outcome_name <- c("ORR", "OS")

res <- extract_aact(con=NULL, path_raw=NULL, terms_int=terms_int, terms_cond=terms_cond,
outcome_list = outcome_list, outcome_name = outcome_name, verbose = TRUE)

}

}
