% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{keywords}
\alias{keywords}
\title{Keywords Dataset}
\format{
A data frame with 145789 rows and 4 variables.
}
\usage{
keywords
}
\description{
Provides words or phrases that best describe the protocol. Keywords help users find studies in the database.
Can include NLM's Medical Subject Heading (MeSH)-controlled vocabulary terms.
}
\details{
This illustrative data only includes participants with conditions "oncology", "malignant neoplasm", and "cancer" based on data extracted on 2022-06-07.

Definition of each variable can be found in the AACT Data Elements table on the AACT Data Dictionary page:
\url{https://aact.ctti-clinicaltrials.org/data_dictionary}
}
\keyword{datasets}
