% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_aact_data.R
\name{load_aact_data}
\alias{load_aact_data}
\title{load_aact_data: Load AACT data into the global R environment}
\usage{
load_aact_data(
  con,
  path_raw,
  board,
  table_names = NULL,
  user_nct_ids = NULL,
  database = NULL,
  env = .GlobalEnv,
  verbose = TRUE
)
}
\arguments{
\item{con}{Connection information for AACT (needed if connecting to database)}

\item{path_raw}{File path for raw AACT data (will ignore "con" argument if provided)}

\item{board}{Pins board}

\item{table_names}{Vector of AACT table-name(s) to load (ex: c("conditions", "interventions") )}

\item{user_nct_ids}{Vector of NCT IDs to filter search on (default=NULL; no filter)}

\item{database}{Only used if con is not NULL. This is only needed if AACT data-tables are
stored such that data-tables are extracted using "select * from database.baseline" etc.}

\item{env}{Environment to save the tables (default=.GlobalEnv)}

\item{verbose}{If TRUE, messages describing process will be printed}
}
\description{
Function that loads AACT data-tables into the R environment. Can be used to pre-load data, which can make
queries a bit faster.
}
