% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{load_data: Load AACT data from local or AACT database}
\usage{
load_data(
  con,
  path_raw,
  file_type,
  board,
  table_name,
  is_shiny_data = NULL,
  user_nct_ids = NULL,
  database = NULL
)
}
\arguments{
\item{con}{Connection information for AACT (needed if connecting to database)}

\item{path_raw}{File path for raw AACT data (will ignore "con" argument if provided)}

\item{file_type}{File type for the raw AACT (ex: "txt" or "rda")}

\item{board}{Pins board}

\item{table_name}{AACT table-name (ex: conditions, interventions)}

\item{is_shiny_data}{Use tables that exist in the global environment. For Shiny apps.}

\item{user_nct_ids}{Vector of NCT IDs to filter search on (default=NULL; no filter)}

\item{database}{Only used if con is not NULL. This is only needed if AACT data-tables are
stored such that data-tables are extracted using "select * from database.baseline" etc.}
}
\value{
AACT data table
}
\description{
Function to load a single AACT data-table and return as data table. This function
is also used internally for the "load_aact_data" function.
}
