% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{outcome_analyses}
\alias{outcome_analyses}
\title{Outcome Analyses Dataset}
\format{
A data frame with 9600 rows and 22 variables.
}
\usage{
outcome_analyses
}
\description{
Results of scientifically appropriate statistical analyses performed on primary and secondary study outcomes. Includes results
for treatment effect estimates, confidence intervals and othe rmeasures of dispersion, and p-values.
}
\details{
This illustrative data only includes participants with conditions "oncology", "malignant neoplasm", and "cancer" based on data extracted on 2022-06-07.

Definition of each variable can be found in the AACT Data Elements table on the AACT Data Dictionary page:
\url{https://aact.ctti-clinicaltrials.org/data_dictionary}
}
\keyword{datasets}
