% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{outcome_analysis_groups}
\alias{outcome_analysis_groups}
\title{Outcome Analysis Dataset}
\format{
A data frame with 18015 rows and 5 variables.
}
\usage{
outcome_analysis_groups
}
\description{
Identifies the comparison groups that were involved with each outcome analysis.
}
\details{
This illustrative data only includes participants with conditions "oncology", "malignant neoplasm", and "cancer" based on data extracted on 2022-06-07.

Definition of each variable can be found in the AACT Data Elements table on the AACT Data Dictionary page:
\url{https://aact.ctti-clinicaltrials.org/data_dictionary}
}
\keyword{datasets}
