% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{outcome_measurements}
\alias{outcome_measurements}
\title{Outcome Measurements Dataset}
\format{
A data frame with 222774 row and 19 variable.
}
\usage{
outcome_measurements
}
\description{
Summary data for primary and secondary outcome measures for each study group.
 Includes parameter estimates and measures of dispersion/precision.
}
\details{
This illustrative data only includes participants with conditions "oncology", "malignant neoplasm", and "cancer" based on data extracted on 2022-06-07.

Definition of each variable can be found in the AACT Data Elements table on the AACT Data Dictionary page:
\url{https://aact.ctti-clinicaltrials.org/data_dictionary}
}
\keyword{datasets}
