% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_outcomes.R
\name{process_outcomes}
\alias{process_outcomes}
\title{process_outcomes: Process outcome measurements / analyses data-tables.}
\usage{
process_outcomes(
  con,
  path_raw,
  file_type,
  board,
  user_nct_ids = NULL,
  database
)
}
\arguments{
\item{con}{Connection information for AACT (needed if connecting to database)}

\item{path_raw}{File path for raw AACT data (will ignore "con" argument if provided)}

\item{file_type}{File type for the raw AACT (ex: "txt" or "rda")}

\item{board}{Pins board}

\item{user_nct_ids}{Vector of unique NCT IDs (default=NULL)}

\item{database}{Only used if con is not NULL. This is only needed if AACT data-tables are
stored such that data-tables are extracted using "select * from database.baseline" etc.}
}
\description{
Function that performs some pre-processing on the outcome measurements and
outcome analyses tables, along with pulling some information from the baseline data-table (work in progress).
}
\seealso{
\code{\link{query_outcomes}}
}
