% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_aact.R
\name{query_aact}
\alias{query_aact}
\title{query_aact: Find trials with intervention(s) and/or condition(s) of interest}
\usage{
query_aact(
  con = NULL,
  path_raw = NULL,
  file_type = NULL,
  board = NULL,
  database = NULL,
  is_shiny = FALSE,
  is_shiny_data = NULL,
  terms_int,
  terms_cond = NULL,
  date_range = NULL,
  related_terms_int = NULL,
  related_terms_cond = NULL,
  verbose = TRUE
)
}
\arguments{
\item{con}{Connection information for AACT (needed if connecting to database)}

\item{path_raw}{File path for raw AACT data (will ignore "con" argument if provided)}

\item{file_type}{File type for the raw AACT (ex: "txt" or "rda")}

\item{board}{Pins board}

\item{database}{Only used if con is not NULL. This is only needed if AACT data-tables are
stored such that data-tables are extracted using "select * from database.baseline" etc.}

\item{is_shiny}{Set to TRUE if used within a SHINY app}

\item{is_shiny_data}{Use tables that exist in the global environment. For Shiny apps.}

\item{terms_int}{Intervention / treatment terms to search for (ex: paclitaxel). Set to NULL if no search.}

\item{terms_cond}{Search terms for condition(s) of interest (ex: non small lung cancer). Set to NULL if no search.}

\item{date_range}{Date range of interest (ex: TBD). Set to NULL if no search.}

\item{related_terms_int}{Related terms for intervention(s) (terms_int).
Default=NULL, which scrapes clinicaltrials.gov for term(s) related to terms_int}

\item{related_terms_cond}{Related terms for condition(s) (terms_cond).
Default=NULL, which scrapes clinicaltrials.gov for term(s) related to terms_cond}

\item{verbose}{If TRUE, messages describing process will be printed}
}
\description{
For intervention (treatment) and/or disease condition search terms, query the AACT database and output
data-set based on the search. By default, for a given search term, this function will web-scrap clinicaltrials.gov to
find related terms (ex: input = "breast cancer", related = "breast tumor").
}
\examples{

\donttest{

library(aactreveal)

# Load package RDA for illustrative purposes #
table_names <- c("browse_interventions", "conditions", "design_group_interventions",
"design_groups","designs", "interventions", "outcome_analyses", "outcome_analysis_groups",
"outcome_measurements", "result_groups", "sponsors", "studies")

data(list=table_names, package="aactreveal")

# Query based on just intervention (any disease indications) #
terms_int <- "pembrolizumab" # Intervention / treatment
out_query1 <- query_aact(terms_int = terms_int, terms_cond = NULL)
out_query1$trt_search_terms # Input search terms + related terms
out_query1$cond_search_terms # Input search terms + related terms

# Query based on intervention and disease condition #
terms_cond <- "breast cancer"
out_query2 <- query_aact(terms_int = terms_int, terms_cond = terms_cond)
out_query2$trt_search_terms # Input search terms + related terms
out_query2$cond_search_terms # Input search terms + related terms


}

}
