% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{result_groups}
\alias{result_groups}
\title{Result Groups Dataset}
\format{
A data frame with 54057 rows and 6 variables.
}
\usage{
result_groups
}
\description{
Consolidated, aggregate list of group titles/descriptions used for reporting summary results information.
}
\details{
This illustrative data only includes participants with conditions "oncology", "malignant neoplasm", and "cancer" based on data extracted on 2022-06-07.

Definition of each variable can be found in the AACT Data Elements table on the AACT Data Dictionary page:
\url{https://aact.ctti-clinicaltrials.org/data_dictionary}
}
\keyword{datasets}
