% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_intervention.R
\name{setup_intervention}
\alias{setup_intervention}
\title{setup_intervention: Find trials with intervention(s) of interest}
\usage{
setup_intervention(
  con,
  path_raw,
  file_type,
  board,
  database,
  is_shiny = FALSE,
  is_shiny_data = NULL,
  search_terms,
  related_terms = NULL
)
}
\arguments{
\item{con}{Connection information for AACT (needed if connecting to database)}

\item{path_raw}{File path for raw AACT data (will ignore "con" argument if provided)}

\item{file_type}{File type for the raw AACT (ex: "txt" or "rda")}

\item{board}{Pins board}

\item{database}{Only used if con is not NULL. This is only needed if AACT data-tables are
stored such that data-tables are extracted using "select * from database.baseline" etc.}

\item{is_shiny}{Set to TRUE if used within a SHINY app. Default=FALSE}

\item{is_shiny_data}{Use tables that exist in the global environment. For Shiny apps.}

\item{search_terms}{Vector of terms to search for (ex: chemotherapy)}

\item{related_terms}{If NULL (default), search for related terms using clinicaltrials.gov.
Otherwise related_terms should be a vector of related terms}
}
\value{
AACT intervention data with treatments of interest
}
\description{
Details here.
Expanding on intervention_setup() by using find_related_terms() (web scraping) to get
related terms from clinicaltrials.gov and putting these terms in str_detect
}
