% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sponsors}
\alias{sponsors}
\title{Sponsors Dataset}
\format{
A data frame with 71652 row and 5 variables.
}
\usage{
sponsors
}
\description{
Name of study sponsors and collaborators. The sponsor is the entity or individual initiating the study. Collaborators are other organizations providing support,
including funding, design, implementation, data analysis, and reporting.
}
\details{
This illustrative data only includes participants with conditions "oncology", "malignant neoplasm", and "cancer" based on data extracted on 2022-06-07.

Definition of each variable can be found in the AACT Data Elements table on the AACT Data Dictionary page:
\url{https://aact.ctti-clinicaltrials.org/data_dictionary}
}
\keyword{datasets}
